src/api/requirements.txt
anyio==3.7.1
certifi==2023.7.22
click==8.1.6
exceptiongroup==1.1.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
numpy==1.24.4
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
sniffio==1.3.0
starlette==0.31.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.2
